refactor(phase8 #66): restore quota routes on v2
