+++ aperag/app.py
@@ -44,6 +44,7 @@ from aperag.domains.conversation.service.bot_service import set_quota_ops as _co
 from aperag.domains.evaluation.api.routes import router as evaluation_v2_router
 from aperag.domains.governance.api.apikeys_routes import router as apikeys_router
 from aperag.domains.governance.api.audit_routes import router as audit_router
+from aperag.domains.governance.api.quota_routes import router as quota_router
 from aperag.domains.identity.api.auth_routes import router as auth_router
 from aperag.domains.identity.api.config_routes import router as config_router
 from aperag.domains.identity.service.user_manager import (
@@ -235,6 +236,7 @@ app.include_router(auth_router, prefix="/api/v2/auth")
 app.include_router(export_router, prefix="/api/v2")  # KB-domain export router (Phase 8 #47 G1, D7 v2 hard-cut)
 app.include_router(audit_router, prefix="/api/v2")  # Governance: audit-logs (hard-cut to v2 in #50)
 app.include_router(apikeys_router, prefix="/api/v2")  # Governance: api_keys (hard-cut to v2 in #51)
+app.include_router(quota_router, prefix="/api/v2")  # Governance: quota/system defaults (hard-cut to v2 in #66)
 app.include_router(llm_router, prefix="/api/v1")  # Model platform: embed/rerank (OpenAI-compat)
 app.include_router(marketplace_router, prefix="/api/v2")  # Marketplace domain router (Phase 8 #52 G4c, D7 v2 hard-cut)
 app.include_router(settings_router, prefix="/api/v2")  # KB domain settings (carved from views/ in #48)
